% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{codings}
\alias{codings}
\title{Example reliability data}
\format{
A matrix with 4 rows and 12 columns. Each column contains the
coders' assessments of a coding unit (e.g. newspaper article)
}
\source{
Krippendorff, K. (1980). Content analysis: An introduction to its
methodology. Beverly Hills, CA: Sage.
}
\usage{
codings
}
\description{
A matrix containing example codings of 12 units (e.g. newspaper articles)
by four coders.
}
\keyword{datasets}
